requirements.txt
playwright==1.40.0
anthropic==0.18.0
asyncpg==0.29.0
pydantic==2.5.0
pyyaml==6.0.1
click==8.1.7
python-dotenv==1.0.0
